update rest-client code.
