+++ src/main/java/xyz/groundx/caver_ext_kas/rest_client/io/swagger/client/api/kip37/model/DeployKip37ContractRequest.java
@@ -21,7 +21,7 @@ import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.io.IOException;
-import xyz.groundx.caver_ext_kas.rest_client.io.swagger.client.api.kip37.model.Kip37FeePayerOption;
+import xyz.groundx.caver_ext_kas.rest_client.io.swagger.client.api.kip37.model.Kip37FeePayerOptions;
 /**
  * DeployKip37ContractRequest
  */
@@ -35,7 +35,7 @@ public class DeployKip37ContractRequest {
   private String uri = null;
 
   @SerializedName("options")
-  private Kip37FeePayerOption options = null;
+  private Kip37FeePayerOptions options = null;
 
   public DeployKip37ContractRequest alias(String alias) {
     this.alias = alias;
@@ -73,7 +73,7 @@ public class DeployKip37ContractRequest {
     this.uri = uri;
   }
 
-  public DeployKip37ContractRequest options(Kip37FeePayerOption options) {
+  public DeployKip37ContractRequest options(Kip37FeePayerOptions options) {
     this.options = options;
     return this;
   }
@@ -83,17 +83,17 @@ public class DeployKip37ContractRequest {
    * @return options
   **/
   @Schema(description = "")
-  public Kip37FeePayerOption getOptions() {
+  public Kip37FeePayerOptions getOptions() {
     return options;
   }
 
-  public void setOptions(Kip37FeePayerOption options) {
+  public void setOptions(Kip37FeePayerOptions options) {
     this.options = options;
   }
 
 
   @Override
-  public boolean equals(Object o) {
+  public boolean equals(java.lang.Object o) {
     if (this == o) {
       return true;
     }
@@ -128,7 +128,7 @@ public class DeployKip37ContractRequest {
    * Convert the given object to string with each line indented by 4 spaces
    * (except the first line).
    */
-  private String toIndentedString(Object o) {
+  private String toIndentedString(java.lang.Object o) {
     if (o == null) {
       return "null";
     }
