+++ src/main/java/xyz/groundx/caver_ext_kas/rest_client/io/swagger/client/api/kip37/model/ImportKip37ContractRequest.java
@@ -21,7 +21,7 @@ import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.io.IOException;
-import xyz.groundx.caver_ext_kas.rest_client.io.swagger.client.api.kip37.model.Kip37FeePayerOption;
+import xyz.groundx.caver_ext_kas.rest_client.io.swagger.client.api.kip37.model.Kip37FeePayerOptions;
 /**
  * ImportKip37ContractRequest
  */
@@ -38,7 +38,7 @@ public class ImportKip37ContractRequest {
   private String address = null;
 
   @SerializedName("options")
-  private Kip37FeePayerOption options = null;
+  private Kip37FeePayerOptions options = null;
 
   public ImportKip37ContractRequest alias(String alias) {
     this.alias = alias;
@@ -94,7 +94,7 @@ public class ImportKip37ContractRequest {
     this.address = address;
   }
 
-  public ImportKip37ContractRequest options(Kip37FeePayerOption options) {
+  public ImportKip37ContractRequest options(Kip37FeePayerOptions options) {
     this.options = options;
     return this;
   }
@@ -104,17 +104,17 @@ public class ImportKip37ContractRequest {
    * @return options
   **/
   @Schema(description = "")
-  public Kip37FeePayerOption getOptions() {
+  public Kip37FeePayerOptions getOptions() {
     return options;
   }
 
-  public void setOptions(Kip37FeePayerOption options) {
+  public void setOptions(Kip37FeePayerOptions options) {
     this.options = options;
   }
 
 
   @Override
-  public boolean equals(Object o) {
+  public boolean equals(java.lang.Object o) {
     if (this == o) {
       return true;
     }
@@ -151,7 +151,7 @@ public class ImportKip37ContractRequest {
    * Convert the given object to string with each line indented by 4 spaces
    * (except the first line).
    */
-  private String toIndentedString(Object o) {
+  private String toIndentedString(java.lang.Object o) {
     if (o == null) {
       return "null";
     }
