+++ src/main/java/xyz/groundx/caver_ext_kas/rest_client/io/swagger/client/api/kip37/model/Kip37FeePayerOptions.java
@@ -21,38 +21,38 @@ import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.io.IOException;
-import xyz.groundx.caver_ext_kas.rest_client.io.swagger.client.api.kip37.model.Kip37FeePayerOptionUserFeePayer;
+import xyz.groundx.caver_ext_kas.rest_client.io.swagger.client.api.kip37.model.Kip37FeePayerOptionsUserFeePayer;
 /**
- * Kip37FeePayerOptionResponse
+ * Kip37FeePayerOptions
  */
 
 
-public class Kip37FeePayerOptionResponse {
-  @SerializedName("enableGlobalFeepayer")
-  private Boolean enableGlobalFeepayer = null;
+public class Kip37FeePayerOptions {
+  @SerializedName("enableGlobalFeePayer")
+  private Boolean enableGlobalFeePayer = null;
 
   @SerializedName("userFeePayer")
-  private Kip37FeePayerOptionUserFeePayer userFeePayer = null;
+  private Kip37FeePayerOptionsUserFeePayer userFeePayer = null;
 
-  public Kip37FeePayerOptionResponse enableGlobalFeepayer(Boolean enableGlobalFeepayer) {
-    this.enableGlobalFeepayer = enableGlobalFeepayer;
+  public Kip37FeePayerOptions enableGlobalFeePayer(Boolean enableGlobalFeePayer) {
+    this.enableGlobalFeePayer = enableGlobalFeePayer;
     return this;
   }
 
    /**
    * A boolean value of whether KAS Global FeePayer is used. The default value is &#x60;true&#x60;.
-   * @return enableGlobalFeepayer
+   * @return enableGlobalFeePayer
   **/
   @Schema(description = "A boolean value of whether KAS Global FeePayer is used. The default value is `true`.")
-  public Boolean isEnableGlobalFeepayer() {
-    return enableGlobalFeepayer;
+  public Boolean isEnableGlobalFeePayer() {
+    return enableGlobalFeePayer;
   }
 
-  public void setEnableGlobalFeepayer(Boolean enableGlobalFeepayer) {
-    this.enableGlobalFeepayer = enableGlobalFeepayer;
+  public void setEnableGlobalFeePayer(Boolean enableGlobalFeePayer) {
+    this.enableGlobalFeePayer = enableGlobalFeePayer;
   }
 
-  public Kip37FeePayerOptionResponse userFeePayer(Kip37FeePayerOptionUserFeePayer userFeePayer) {
+  public Kip37FeePayerOptions userFeePayer(Kip37FeePayerOptionsUserFeePayer userFeePayer) {
     this.userFeePayer = userFeePayer;
     return this;
   }
@@ -62,40 +62,40 @@ public class Kip37FeePayerOptionResponse {
    * @return userFeePayer
   **/
   @Schema(description = "")
-  public Kip37FeePayerOptionUserFeePayer getUserFeePayer() {
+  public Kip37FeePayerOptionsUserFeePayer getUserFeePayer() {
     return userFeePayer;
   }
 
-  public void setUserFeePayer(Kip37FeePayerOptionUserFeePayer userFeePayer) {
+  public void setUserFeePayer(Kip37FeePayerOptionsUserFeePayer userFeePayer) {
     this.userFeePayer = userFeePayer;
   }
 
 
   @Override
-  public boolean equals(Object o) {
+  public boolean equals(java.lang.Object o) {
     if (this == o) {
       return true;
     }
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    Kip37FeePayerOptionResponse kip37FeePayerOptionResponse = (Kip37FeePayerOptionResponse) o;
-    return Objects.equals(this.enableGlobalFeepayer, kip37FeePayerOptionResponse.enableGlobalFeepayer) &&
-        Objects.equals(this.userFeePayer, kip37FeePayerOptionResponse.userFeePayer);
+    Kip37FeePayerOptions kip37FeePayerOptions = (Kip37FeePayerOptions) o;
+    return Objects.equals(this.enableGlobalFeePayer, kip37FeePayerOptions.enableGlobalFeePayer) &&
+        Objects.equals(this.userFeePayer, kip37FeePayerOptions.userFeePayer);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(enableGlobalFeepayer, userFeePayer);
+    return Objects.hash(enableGlobalFeePayer, userFeePayer);
   }
 
 
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    sb.append("class Kip37FeePayerOptionResponse {\n");
+    sb.append("class Kip37FeePayerOptions {\n");
     
-    sb.append("    enableGlobalFeepayer: ").append(toIndentedString(enableGlobalFeepayer)).append("\n");
+    sb.append("    enableGlobalFeePayer: ").append(toIndentedString(enableGlobalFeePayer)).append("\n");
     sb.append("    userFeePayer: ").append(toIndentedString(userFeePayer)).append("\n");
     sb.append("}");
     return sb.toString();
@@ -105,7 +105,7 @@ public class Kip37FeePayerOptionResponse {
    * Convert the given object to string with each line indented by 4 spaces
    * (except the first line).
    */
-  private String toIndentedString(Object o) {
+  private String toIndentedString(java.lang.Object o) {
     if (o == null) {
       return "null";
     }
