+++ src/main/java/xyz/groundx/caver_ext_kas/rest_client/io/swagger/client/api/kip37/model/Kip37TransactionStatusResponse.java
@@ -92,7 +92,7 @@ public class Kip37TransactionStatusResponse {
 
 
   @Override
-  public boolean equals(Object o) {
+  public boolean equals(java.lang.Object o) {
     if (this == o) {
       return true;
     }
@@ -127,7 +127,7 @@ public class Kip37TransactionStatusResponse {
    * Convert the given object to string with each line indented by 4 spaces
    * (except the first line).
    */
-  private String toIndentedString(Object o) {
+  private String toIndentedString(java.lang.Object o) {
     if (o == null) {
       return "null";
     }
